Rename `classDialogBox.py` -> `dialog_box.py`.
